# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TalentPool API Documentation",
        "description": "API documentation for TalentPool application",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/annonces": {
            "get": {
                "tags": [
                    "Annonces"
                ],
                "summary": "Get all announcements",
                "description": "Display a listing of all announcements",
                "operationId": "31d0e6c7ff05a37add8b40190cb07b88",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Annonce"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Annonces"
                ],
                "summary": "Create new announcement",
                "description": "Store a newly created announcement",
                "operationId": "6d7f668673a408bd0425d541f43dd9d8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Annonce"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Announcement created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Annonce"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/annonces/{id}": {
            "get": {
                "tags": [
                    "Annonces"
                ],
                "summary": "Get announcement by ID",
                "description": "Display the specified announcement",
                "operationId": "866cca6c534405197636d5b87cbb01dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Annonce"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Announcement not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Annonces"
                ],
                "summary": "Update announcement",
                "description": "Update the specified announcement",
                "operationId": "9e5ad10772a81ce5556d61802538dad5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Annonce"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Announcement updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Annonce"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Announcement not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Annonces"
                ],
                "summary": "Delete announcement",
                "description": "Remove the specified announcement",
                "operationId": "ff957a7f8f8f88e3747c6f059eda335a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Announcement deleted successfully"
                    },
                    "404": {
                        "description": "Announcement not found"
                    }
                }
            }
        },
        "/api/candidatures": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Get all candidatures",
                "operationId": "4433f2edc959ef215a18a81309e913df",
                "responses": {
                    "200": {
                        "description": "List of candidatures",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Candidature"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Create new candidature",
                "operationId": "62704344c7a43c99304e8d0a1c5e956e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Candidature"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidature created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Candidature"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/candidatures/{id}": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Get candidature by ID",
                "operationId": "b4424ba45630d3949a32547388c529b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Candidature details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Candidature"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Candidature not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Update candidature",
                "operationId": "048e7854d052736e6a721b6301427f5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Candidature"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidature updated successfully"
                    },
                    "404": {
                        "description": "Candidature not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Delete candidature",
                "operationId": "2eed460642eada7d704c958dabac6ddd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Candidature deleted successfully"
                    },
                    "404": {
                        "description": "Candidature not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Annonce": {
                "title": "Annonce",
                "description": "Annonce model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "recruteur_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "titre": {
                        "type": "string",
                        "example": "Software Developer Position"
                    },
                    "description": {
                        "type": "string",
                        "example": "We are looking for a talented developer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Candidature": {
                "title": "Candidature",
                "description": "Candidature model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "candidat_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "annonce_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "cv": {
                        "type": "string",
                        "example": "path/to/cv.pdf"
                    },
                    "lettre_motivation": {
                        "type": "string",
                        "example": "path/to/letter.pdf"
                    },
                    "statu": {
                        "type": "string",
                        "example": "pending"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Annonces",
            "description": "Annonces"
        },
        {
            "name": "Candidatures",
            "description": "Candidatures"
        }
    ]
}